Automate the creation of new PMO site repo
This creates a new repository on GitHub for a PMO site and creates two
initial issues, one for technical issues and one for non-technical
issues.

The idea here is that we can slowly automate away each item on the
checklist so that it can be done as part of the initial setup.